Stronger model, more efficient sentiment calculations
Increased the strength of the model by using 15,000 reviews for each rating (1-5 stars) instead of 50,000 reviews of random score (which heavily weighted towards 5 stars and 1 star)

Reviews are now sent to the model for sentiment analysis in batches of 50 instead of all at once. This avoids overloading the GPU's memory and results in dramatically faster processing.